+++ neo-cli/CLI/MainService.cs
@@ -443,186 +443,6 @@ namespace Neo.CLI
             return output;
         }
 
-        public static string ToCLIString(Cosigner cosigner)
-        {
-            string output = $"\tAccount: {cosigner.Account}\n";
-
-            output += "\tScope:\t";
-            if (cosigner.Scopes == WitnessScope.Global)
-            {
-                output += $"Global\t";
-            }
-            if (cosigner.Scopes.HasFlag(WitnessScope.CalledByEntry))
-            {
-                output += $"CalledByEntry\t";
-            }
-            if (cosigner.Scopes.HasFlag(WitnessScope.CustomContracts))
-            {
-                output += $"CustomContract\t";
-            }
-            if (cosigner.Scopes.HasFlag(WitnessScope.CustomGroups))
-            {
-                output += $"CustomGroup\t";
-            }
-
-            output += "\n";
-
-            if (cosigner.AllowedContracts != null && cosigner.AllowedContracts.Length > 0)
-            {
-                output += "Allowed contracts: \n";
-                foreach (var allowedContract in cosigner.AllowedContracts)
-                {
-                    output += $"\t{allowedContract.ToString()}\n";
-                }
-            }
-
-            if (cosigner.AllowedGroups != null && cosigner.AllowedGroups.Length > 0)
-            {
-                output += "Allowed groups: \n";
-                foreach (var allowedGroup in cosigner.AllowedGroups)
-                {
-                    output += $"\t{allowedGroup.ToString()}\n";
-                }
-            }
-
-            return output;
-        }
-
-        public static string ToCLIString(Witness witness)
-        {
-            string output = "";
-            output += $"Invocation: \n";
-            output += DisassembleScript(witness.InvocationScript);
-            output += $"Verification: \n";
-            output += DisassembleScript(witness.VerificationScript);
-            return output;
-        }
-
-
-        public static string ToCLIString(NotifyEventArgs notification)
-        {
-            string output = "";
-            var vmArray = (Neo.VM.Types.Array)notification.State;
-            var notificationName = "";
-            if (vmArray.Count > 0)
-            {
-                notificationName = vmArray[0].GetString();
-            }
-
-            var adapter = GetCliStringAdapter(notificationName);
-            output += adapter(vmArray);
-
-            return output;
-        }
-
-
-        public static string ToCLIString(Block block)
-        {
-            string output = "";
-            output += $"Hash: {block.Hash}\n";
-            output += $"Index: {block.Index}\n";
-            output += $"Size: {block.Size}\n";
-            output += $"PreviousBlockHash: {block.PrevHash}\n";
-            output += $"MerkleRoot: {block.MerkleRoot}\n";
-            output += $"Time: {block.Timestamp}\n";
-            output += $"NextConsensus: {block.NextConsensus}\n";
-            output += $"Transactions:\n";
-            foreach (Transaction t in block.Transactions)
-            {
-                output += $"{ToCLIString(t, block.Timestamp)}";
-                output += $"\n";
-            }
-            output += $"Witnesses:\n";
-            output += $"\tInvocation: {block.Witness.InvocationScript.ToHexString()}\n";
-            output += $"\tVerification: {block.Witness.VerificationScript.ToHexString()}\n";
-            return output;
-        }
-
-        public static string ToCLIString(Transaction t, ulong blockTimestamp = 0)
-        {
-            string output = "";
-            output += $"Hash: {t.Hash}\n";
-            if (blockTimestamp > 0)
-            {
-                var blockTime = UnixEpoch.AddMilliseconds(blockTimestamp);
-                blockTime = TimeZoneInfo.ConvertTimeFromUtc(blockTime, TimeZoneInfo.Local);
-                output += $"Timestamp: {blockTime.ToShortDateString()} {blockTime.ToLongTimeString()}\n";
-            }
-            output += $"NetFee: {new BigDecimal(t.NetworkFee, NeoToken.GAS.Decimals)}\n";
-            output += $"SysFee: {new BigDecimal(t.SystemFee, NeoToken.GAS.Decimals)}\n";
-            output += $"Sender: {t.Sender.ToAddress()}\n";
-            if (t.Cosigners != null && t.Cosigners.Length > 0)
-            {
-                output += $"Cosigners:\n";
-                foreach (var cosigner in t.Cosigners)
-                {
-                    output += ToCLIString(cosigner);
-                }
-            }
-
-
-            output += $"Script:\n";
-            output += DisassembleScript(t.Script);
-
-            output += "Witnesses: \n";
-            foreach (var witness in t.Witnesses)
-            {
-                output += $"{ToCLIString(witness)}";
-            }
-
-
-            return output;
-        }
-
-
-        public static string ToCLIString(ContractState c)
-        {
-            string output = "";
-            output += $"Hash: {c.ScriptHash}\n";
-            output += $"EntryPoint: \n";
-            output += $"\tName: {c.Manifest.Abi.EntryPoint.Name}\n";
-            output += $"\tParameters: \n";
-            foreach (var parameter in c.Manifest.Abi.EntryPoint.Parameters)
-            {
-                output += $"\t\tName: {parameter.Name}\n";
-                output += $"\t\tType: {parameter.Type}\n";
-            }
-
-            output += $"Methods: \n";
-            foreach (var method in c.Manifest.Abi.Methods)
-            {
-                output += $"\tName: {method.Name}\n";
-                output += $"\tReturn Type: {method.ReturnType}\n";
-                if (method.Parameters.Length > 0)
-                {
-                    output += $"\tParameters: \n";
-                    foreach (var parameter in method.Parameters)
-                    {
-                        output += $"\t\tName: {parameter.Name}\n";
-                        output += $"\t\tType: {parameter.Type}\n";
-                    }
-                }
-
-            }
-
-            output += $"Events: \n";
-            foreach (var abiEvent in c.Manifest.Abi.Events)
-            {
-                output += $"\tName: {abiEvent.Name}\n";
-                if (abiEvent.Parameters.Length > 0)
-                {
-                    output += $"\tParameters: \n";
-                    foreach (var parameter in abiEvent.Parameters)
-                    {
-                        output += $"\t\tName: {parameter.Name}\n";
-                        output += $"\t\tType: {parameter.Type}\n";
-                    }
-                }
-            }
-
-            return output;
-        }
-
         private static Dictionary<string, Func<VM.Types.Array, string>> notificationAdapters = new Dictionary<string, Func<VM.Types.Array, string>>()
         {
             {"transfer", TransferNotificationCLIStringAdapter},
